Split jwt directory into jwt/core and jwt/http
