+++ packages/jwt/src/core/index.ts
@@ -1,0 +1,2 @@
+export { getSecretOrPrivateKey } from './get-secret-or-private-key.util';
+export { getSecretOrPublicKey } from './get-secret-or-public-key.util';
