+++ packages/jwt/src/http/index.ts
@@ -1,0 +1,19 @@
+export {
+  InvalidTokenError,
+  isInvalidTokenError,
+} from './invalid-token.error';
+export {
+  JWTOptional,
+} from './jwt-optional.hook';
+export {
+  JWTRequired,
+} from './jwt-required.hook';
+export {
+  JWTOptions, VerifyOptions,
+} from './jwt.hook';
+export {
+  removeAuthCookie,
+} from './remove-auth-cookie';
+export {
+  setAuthCookie,
+} from './set-auth-cookie';
