+++ packages/jwt/src/http/jwt-optional.hook.spec.ts
