+++ packages/jwt/src/http/jwt.hook.ts
@@ -19,7 +19,7 @@ import { decode, verify } from 'jsonwebtoken';
 
 // FoalTS
 import { JWT_DEFAULT_COOKIE_NAME, JWT_DEFAULT_CSRF_COOKIE_NAME } from './constants';
-import { getSecretOrPublicKey } from './get-secret-or-public-key.util';
+import { getSecretOrPublicKey } from '../core';
 import { isInvalidTokenError } from './invalid-token.error';
 
 class InvalidTokenResponse extends HttpResponseUnauthorized {
