+++ packages/jwt/src/http/remove-auth-cookie.spec.ts
