+++ packages/jwt/src/http/remove-auth-cookie.ts
