+++ packages/jwt/src/index.ts
@@ -6,26 +6,14 @@
 
 export {
   getSecretOrPrivateKey,
-} from './get-secret-or-private-key.util';
-export {
   getSecretOrPublicKey,
-} from './get-secret-or-public-key.util';
+} from './core';
 export {
   InvalidTokenError,
   isInvalidTokenError,
-} from './invalid-token.error';
-export {
   JWTOptional,
-} from './jwt-optional.hook';
-export {
   JWTRequired,
-} from './jwt-required.hook';
-export {
   JWTOptions, VerifyOptions,
-} from './jwt.hook';
-export {
   removeAuthCookie,
-} from './remove-auth-cookie';
-export {
   setAuthCookie,
-} from './set-auth-cookie';
+} from './http';
